Update Dockerfile documentation on build cache
I propose to add some line in the documentation on how to use a Dockerfile after running in build cache problems, see #173. As a new user of mybinder, the difference the treatment of the Dockerfile by dockerhub and Binderhub was very confusing, and I hope this (or any better/more concise formulation) might help users with similar problems.